requirements.txt
openai>=1.0.0
matplotlib
pandas
reportlab
jinja2
python-dotenv
